+++ restfmclient/cursor.py
@@ -8,7 +8,7 @@ import asyncio
 
 
 class Cursor(AsyncIterator):
-    def __init__(self, client, block_size=100,
+    def __init__(self, client, block_size=128,
                  limit=None, offset=0, prefetch=True, field_info=None):
         self._client = client
         self._limit = limit