+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 ---------
 
-1.1.1 (unreleased)
+1.2.0 (2016-12-17)
 ------------------
 
 - Improve tests, add more Date types.